+++ activestorage/test/models/attachment_test.rb
@@ -100,6 +100,23 @@ class ActiveStorage::AttachmentTest < ActiveSupport::TestCase
     assert_equal blob, ActiveStorage::Blob.find_signed!(signed_id, purpose: :custom_purpose)
   end
 
+  test "getting a signed blob ID from an attachment with a expires_in" do
+    blob = create_blob
+    @user.avatar.attach(blob)
+
+    signed_id = @user.avatar.signed_id(expires_in: 1.minute)
+    assert_equal blob, ActiveStorage::Blob.find_signed!(signed_id)
+  end
+
+  test "fail to find blob within expiration date" do
+    blob = create_blob
+    @user.avatar.attach(blob)
+
+    signed_id = @user.avatar.signed_id(expires_in: 1.minute)
+    travel 2.minutes
+    assert_nil ActiveStorage::Blob.find_signed(signed_id)
+  end
+
   test "signed blob ID backwards compatibility" do
     blob = create_blob
     @user.avatar.attach(blob)